banking-app/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Api Rest com FastAPI da aplicação banking-app"
authors = [
    {name = "Gabriel Dantas",email = "gabrieldantaslp@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "databases",
    "pyjwt",
    "sqlalchemy",
    "alembic",
    "psycopg2[binary]",
    "aiosqlite"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
